feat: filtered comp groups changes
