+++ src/modules/players/forms/filter.tsx
@@ -1,17 +1,18 @@
 import { Box, TextField } from '@mui/material'
-import { Field, Form, Formik, FormikProps } from 'formik'
+import { Field, Form, Formik } from 'formik'
 import { CheckboxWithLabel } from 'formik-mui'
 import { useTranslation } from 'next-i18next'
-import { useRef } from 'react'
 
 import { FilterCombo } from '@/components/combo/combo'
 import { mapListDataToComboOptions } from '@/components/combo/utils'
-import FilteredCompetitionGroups from '@/components/filteredCompetitionGroups/filteredCompetitionGroups'
+import FilteredCompetitonGroups from '@/components/filteredCompetitionGroups/filteredCompetitonGroups'
 import { FilterCheckboxContainer } from '@/components/forms/filter-checkbox-container'
 import { FilterFormActions } from '@/components/forms/filter-form-actions'
 import { FilterFormContainer } from '@/components/forms/filter-form-container'
 import { CompetitionGroupBasicDataDto } from '@/modules/competition-groups/types'
+import { mapCompetitionGroupsListToComboOptions } from '@/modules/competition-groups/utils'
 import { CompetitionBasicDataDto } from '@/modules/competitions/types'
+import { mapCompetitionsListToComboOptions } from '@/modules/competitions/utils'
 import { CountryDto } from '@/modules/countries/types'
 import { mapCountriesListToComboOptions } from '@/modules/countries/utils'
 import { PlayerPositionDto } from '@/modules/player-positions/types'
@@ -44,11 +45,12 @@ export const PlayersFilterForm = ({
 }: IPlayersFilterFormProps) => {
   const { t } = useTranslation(['common', 'players'])
   const footedComboData = getFootedComboData(t)
-  const formRef = useRef<FormikProps<PlayersFiltersState>>(null)
+  const groupsComboData = mapCompetitionGroupsListToComboOptions(
+    competitionGroupsData,
+  )
 
   return (
     <Formik
-      innerRef={formRef}
       initialValues={filters}
       onSubmit={(data, form) => {
         onFilter(data)
@@ -56,7 +58,7 @@ export const PlayersFilterForm = ({
       }}
       enableReinitialize
     >
-      {() => (
+      {({ values }) => (
         <Form autoComplete="off">
           <FilterFormContainer>
             <Field
@@ -116,12 +118,20 @@ export const PlayersFilterForm = ({
               size="small"
               multiple
             />
-            <FilteredCompetitionGroups
-              competitionsFormValues={formRef.current?.values.competitionIds}
-              competitionGroupsData={competitionGroupsData}
-              competitionsData={competitionsData}
-              competitionsMultiple
-              groupsMultiple
+            <FilterCombo
+              data={mapCompetitionsListToComboOptions(competitionsData)}
+              name="competitionIds"
+              label={t('COMPETITIONS')}
+              size="small"
+              multiple
+            />
+            <FilteredCompetitonGroups
+              competitionGroupsData={groupsComboData}
+              competitionsFormValues={values.competitionIds}
+              name="competitionGroupsIds"
+              label={t('COMPETITIONS')}
+              size="small"
+              multiple
             />
           </FilterFormContainer>
           <FilterCheckboxContainer>
